refactor(runtime): move the readiness guard next to the API it wraps
Review follow-ups, all layering and wording — no behavior change.

`ReadinessHold` was a pure RAII wrapper around `hold_endpoint_readiness` /
`release_endpoint_readiness`, living in backend-common but carrying a
lock-reentrancy caveat that is a property of `SystemHealth`'s own locking. Any
other endpoint owner needing the same take-before-register discipline would have
had to reimplement it. Move it beside `SystemHealth` and make it public.

`take` no longer accepts an endpoint plus a separately-passed name. With the
guard scoped to the runtime it takes the health handle and the endpoint name
directly, so there is no second argument that must match the first by
convention.

Align the guard's doc with `hold_endpoint_readiness`: both now state the same
two effects, rather than the guard describing only the per-endpoint one.

Rename `a_hold_does_not_leak_to_other_endpoints` and rewrite its comment. It
claimed "the hold gates only the endpoint it names", which stopped being true
when a hold started suppressing the whole process's readiness. The assertion
itself was always correct — per-endpoint suppression inside
`set_endpoint_registered` really does stay scoped — so only the wording changes.
